shared auth helpers in conftest
