+++ tests/test_auth.py
@@ -1,82 +1,9 @@
 import importlib
-from contextlib import contextmanager
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
-import pytest
 from fastapi.testclient import TestClient
 
-
-def _create_app_with_auth(
-    monkeypatch, introspection_url, client_id="test_client", client_secret="test_secret"
-):
-    """Create app with authentication enabled."""
-    monkeypatch.setenv("AUTH_TOKEN_INTROSPECTION_URL", introspection_url)
-    monkeypatch.setenv("AUTH_CLIENT_ID", client_id)
-    monkeypatch.setenv("AUTH_CLIENT_SECRET", client_secret)
-
-    # reload modules to pick up new env vars
-    import sc_keeper.api
-    import sc_keeper.auth
-
-    importlib.reload(sc_keeper.auth)
-    importlib.reload(sc_keeper.api)
-
-    return sc_keeper.api.app
-
-
-def _create_mock_introspection_response(user_data):
-    """Create a mock token introspection response."""
-    mock_response = Mock()
-    mock_response.json.return_value = user_data
-    mock_response.status_code = 200
-    mock_response.raise_for_status = Mock()
-    return mock_response
-
-
-@contextmanager
-def mock_token_introspection(user_data=None, exception=None):
-    """Context manager to mock token introspection API calls.
-
-    Args:
-        user_data: Dict with token introspection response data (e.g., {"sub": "user123"})
-        exception: Exception to raise instead of returning a response
-
-    Yields:
-        The mock async client (useful for checking call counts)
-    """
-    if exception:
-        mock_response = None
-    else:
-        mock_response = _create_mock_introspection_response(user_data or {})
-
-    class MockAsyncClient:
-        def __init__(self):
-            self.post_call_count = 0
-
-        async def __aenter__(self):
-            return self
-
-        async def __aexit__(self, *args):
-            return None
-
-        async def post(self, *args, **kwargs):
-            self.post_call_count += 1
-            if exception:
-                raise exception
-            return mock_response
-
-    mock_client = MockAsyncClient()
-
-    with patch("sc_keeper.auth.httpx.AsyncClient", return_value=mock_client):
-        yield mock_client
-
-
-@pytest.fixture
-def client_with_auth(monkeypatch):
-    """Create a test client with authentication enabled."""
-    introspection_url = "http://test-auth-server.com/introspect"
-    app = _create_app_with_auth(monkeypatch, introspection_url)
-    return TestClient(app), introspection_url
+from conftest import create_app_with_auth, mock_token_introspection
 
 
 def test_endpoints_no_token(client_with_auth):
@@ -129,7 +56,7 @@ def test_auth_token_missing_scope(monkeypatch):
     """Test that requests with token missing required scope return 401."""
     introspection_url = "http://test-auth-server.com/introspect"
     monkeypatch.setenv("AUTH_TOKEN_VALIDATION_CEL", "claims.scope == 'required_scope'")
-    app = _create_app_with_auth(monkeypatch, introspection_url)
+    app = create_app_with_auth(monkeypatch, introspection_url)
     client = TestClient(app)
 
     with mock_token_introspection(
@@ -149,7 +76,7 @@ def test_auth_token_with_required_scope(monkeypatch):
     """Test that requests with token having required scope succeed."""
     introspection_url = "http://test-auth-server.com/introspect"
     monkeypatch.setenv("AUTH_TOKEN_VALIDATION_CEL", "claims.scope == 'required_scope'")
-    app = _create_app_with_auth(monkeypatch, introspection_url)
+    app = create_app_with_auth(monkeypatch, introspection_url)
     client = TestClient(app)
 
     with mock_token_introspection(
